+++ obj/src/index.d.ts
@@ -1,3 +1,4 @@
 /** @module index */
 export * from './build';
+export * from './connect';
 export * from './queues';