+++ obj/src/queues/LockedMessage.js
@@ -1,5 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.LockedMessage = void 0;
 /**
  * Data object used to store and lock incoming messages
  * in [[MemoryMessageQueue]].