+++ obj/src/queues/MessageQueue.js
@@ -1,13 +1,16 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.MessageQueue = void 0;
 /** @module queues */
 /** @hidden */
-var async = require('async');
+const async = require('async');
 const pip_services3_commons_node_1 = require("pip-services3-commons-node");
+const pip_services3_commons_node_2 = require("pip-services3-commons-node");
 const pip_services3_components_node_1 = require("pip-services3-components-node");
 const pip_services3_components_node_2 = require("pip-services3-components-node");
 const pip_services3_components_node_3 = require("pip-services3-components-node");
 const pip_services3_components_node_4 = require("pip-services3-components-node");
+const MessagingCapabilities_1 = require("./MessagingCapabilities");
 const MessageEnvelope_1 = require("./MessageEnvelope");
 /**
  * Abstract message queue that is used as a basis for specific message queue implementations.
@@ -30,23 +33,26 @@ const MessageEnvelope_1 = require("./MessageEnvelope");
  *
  * ### References ###
  *
- * - <code>\*:logger:\*:\*:1.0</code>           (optional) [[https://rawgit.com/pip-services-node/pip-services3-components-node/master/doc/api/interfaces/log.ilogger.html ILogger]] components to pass log messages
- * - <code>\*:counters:\*:\*:1.0</code>         (optional) [[https://rawgit.com/pip-services-node/pip-services3-components-node/master/doc/api/interfaces/count.icounters.html ICounters]] components to pass collected measurements
- * - <code>\*:discovery:\*:\*:1.0</code>        (optional) [[https://rawgit.com/pip-services-node/pip-services3-components-node/master/doc/api/interfaces/connect.idiscovery.html IDiscovery]] components to discover connection(s)
- * - <code>\*:credential-store:\*:\*:1.0</code> (optional) [[https://rawgit.com/pip-services-node/pip-services3-components-node/master/doc/api/interfaces/auth.icredentialstore.html ICredentialStore]] componetns to lookup credential(s)
+ * - <code>\*:logger:\*:\*:1.0</code>           (optional) [[https://pip-services3-node.github.io/pip-services3-components-node/interfaces/log.ilogger.html ILogger]] components to pass log messages
+ * - <code>\*:counters:\*:\*:1.0</code>         (optional) [[https://pip-services3-node.github.io/pip-services3-components-node/interfaces/count.icounters.html ICounters]] components to pass collected measurements
+ * - <code>\*:discovery:\*:\*:1.0</code>        (optional) [[https://pip-services3-node.github.io/pip-services3-components-node/interfaces/connect.idiscovery.html IDiscovery]] components to discover connection(s)
+ * - <code>\*:credential-store:\*:\*:1.0</code> (optional) [[https://pip-services3-node.github.io/pip-services3-components-node/interfaces/auth.icredentialstore.html ICredentialStore]] componetns to lookup credential(s)
  */
 class MessageQueue {
     /**
      * Creates a new instance of the message queue.
      *
      * @param name  (optional) a queue name
+     * @param capabilities (optional) a capabilities of this message queue
      */
-    constructor(name) {
+    constructor(name, capabilities) {
         this._logger = new pip_services3_components_node_1.CompositeLogger();
         this._counters = new pip_services3_components_node_2.CompositeCounters();
         this._connectionResolver = new pip_services3_components_node_3.ConnectionResolver();
         this._credentialResolver = new pip_services3_components_node_4.CredentialResolver();
         this._name = name;
+        this._capabilities = capabilities
+            || new MessagingCapabilities_1.MessagingCapabilities(false, false, false, false, false, false, false, false, false);
     }
     /**
      * Gets the queue name
@@ -66,7 +72,7 @@ class MessageQueue {
      * @param config    configuration parameters to be set.
      */
     configure(config) {
-        this._name = pip_services3_commons_node_1.NameResolver.resolve(config, this._name);
+        this._name = pip_services3_commons_node_2.NameResolver.resolve(config, this._name);
         this._logger.configure(config);
         this._connectionResolver.configure(config);
         this._credentialResolver.configure(config);
@@ -108,8 +114,19 @@ class MessageQueue {
         this.openWithParams(correlationId, connection, credential, callback);
     }
     /**
+     * Checks if the queue has been opened
+     * @param correlationId     (optional) transaction id to trace execution through call chain.
+     * @returns Error if queue wasn't opened or <code>null</code> otherwise
+     */
+    checkOpen(correlationId) {
+        if (!this.isOpen()) {
+            return new pip_services3_commons_node_1.InvalidStateException(correlationId, "NOT_OPENED", "The queue is not opened");
+        }
+        return null;
+    }
+    /**
      * Sends an object into the queue.
-     * Before sending the object is converted into JSON string and wrapped in a [[MessageEnvelop]].
+     * Before sending the object is converted into JSON string and wrapped in a [[MessageEnvelope]].
      *
      * @param correlationId     (optional) transaction id to trace execution through call chain.
      * @param messageType       a message type