+++ obj/src/queues/index.js
@@ -1,11 +1,14 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.CallbackMessageReceiver = exports.MemoryMessageQueue = exports.MessageQueue = exports.MessagingCapabilities = exports.MessageEnvelope = void 0;
 var MessageEnvelope_1 = require("./MessageEnvelope");
-exports.MessageEnvelope = MessageEnvelope_1.MessageEnvelope;
+Object.defineProperty(exports, "MessageEnvelope", { enumerable: true, get: function () { return MessageEnvelope_1.MessageEnvelope; } });
 var MessagingCapabilities_1 = require("./MessagingCapabilities");
-exports.MessagingCapabilities = MessagingCapabilities_1.MessagingCapabilities;
+Object.defineProperty(exports, "MessagingCapabilities", { enumerable: true, get: function () { return MessagingCapabilities_1.MessagingCapabilities; } });
 var MessageQueue_1 = require("./MessageQueue");
-exports.MessageQueue = MessageQueue_1.MessageQueue;
+Object.defineProperty(exports, "MessageQueue", { enumerable: true, get: function () { return MessageQueue_1.MessageQueue; } });
 var MemoryMessageQueue_1 = require("./MemoryMessageQueue");
-exports.MemoryMessageQueue = MemoryMessageQueue_1.MemoryMessageQueue;
+Object.defineProperty(exports, "MemoryMessageQueue", { enumerable: true, get: function () { return MemoryMessageQueue_1.MemoryMessageQueue; } });
+var CallbackMessageReceiver_1 = require("./CallbackMessageReceiver");
+Object.defineProperty(exports, "CallbackMessageReceiver", { enumerable: true, get: function () { return CallbackMessageReceiver_1.CallbackMessageReceiver; } });
 //# sourceMappingURL=index.js.map