+++ test/queues/MessageQueueFixture.ts
@@ -1,5 +1,5 @@
 let assert = require('chai').assert;
-let async = require('async');
+const async = require('async');
 
 import { IMessageQueue } from '../../src/queues/IMessageQueue';
 import { MessageEnvelope } from '../../src/queues/MessageEnvelope';
@@ -21,12 +21,12 @@ export class MessageQueueFixture {
             (callback) => {
                 this._queue.send(null, envelope1, callback);
             },
-            (callback) => {
-                var count = this._queue.readMessageCount((err, count) => {
-                    assert.isTrue(count > 0);
-                    callback(err);
-                });
-            },
+            // (callback) => {
+            //     var count = this._queue.readMessageCount((err, count) => {
+            //         assert.isTrue(count > 0);
+            //         callback(err);
+            //     });
+            // },
             (callback) => {
                 this._queue.receive(null, 10000, (err, result) => {
                     envelope2 = result;