+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Asynchronous Messaging for Node.js Changelog
 
+## <a name="3.1.0"></a> 3.1.0 (2021-03-23)
+
+Improved message queuues
+
+### Features
+* **queues** Added CallbackMessageReceiver to wrap callbacks into IMessageReceiver interface
+* **queues** Addded IMessageConnection interface
+* **build** Set config params and references to created queues in MessageQueueFactory
+* **queues** Added checkOpen method to MessageQueue
+
 ## <a name="3.0.0"></a> 3.0.0 (2018-11-03)
 
 Initial public release